feat: initial implementation of output
TODO: Coordinate with group to determine structs to be used and how
calculations are to be done. Some outputs are currently not fully
functional as they depend on calculations. As it stands we'd have to do
the calculations _in_ the actual output functions which is a
non-starter. What should be done is being a little more open with what
is explicitly injected into the structs (e.g. for Phys Mem. also require
t number of valid bits or something like that). Has to be determined by
the group.